Move BoardItem into its on memoized component.
